+++ src/components/PatronAuthServices.tsx
@@ -33,7 +33,7 @@ function mapStateToProps(state, ownProps) {
   }
   return {
     data: data,
-    responseBody: state.editor.patronAuthServices && state.editor.patronAuthServices.responseBody,
+    responseBody: state.editor.patronAuthServices && state.editor.patronAuthServices.successMessage,
     fetchError: state.editor.patronAuthServices.fetchError,
     isFetching: state.editor.patronAuthServices.isFetching || state.editor.patronAuthServices.isEditing
   };