+++ src/components/SearchServices.tsx
@@ -21,7 +21,7 @@ function mapStateToProps(state, ownProps) {
   const data = Object.assign({}, state.editor.searchServices && state.editor.searchServices.data || {});
   return {
     data: data,
-    responseBody: state.editor.searchServices && state.editor.searchServices.responseBody,
+    responseBody: state.editor.searchServices && state.editor.searchServices.successMessage,
     fetchError: state.editor.searchServices.fetchError,
     isFetching: state.editor.searchServices.isFetching || state.editor.searchServices.isEditing
   };