+++ src/components/ServiceEditForm.tsx
@@ -1,17 +1,20 @@
 import * as React from "react";
 import EditableInput from "./EditableInput";
 import ProtocolFormField from "./ProtocolFormField";
+import SaveButton from "./SaveButton";
+import Collapsible from "./Collapsible";
 import WithEditButton from "./WithEditButton";
 import WithRemoveButton from "./WithRemoveButton";
 import { LibraryData, LibraryWithSettingsData, ProtocolData, ServiceData, ServicesData } from "../interfaces";
 import { EditFormProps } from "./EditableConfigList";
+import { handleSubmit, clearForm } from "./sharedFunctions";
 import { FetchErrorData } from "opds-web-client/lib/interfaces";
 
 export interface ServiceEditFormProps<T> {
   data: T;
   item?: ServiceData;
   disabled: boolean;
-  editItem: (data: FormData) => Promise<void>;
+  save: (data: FormData) => void;
   urlBase: string;
   listDataKey: string;
   responseBody?: string;
@@ -49,12 +52,13 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
     this.editLibrary = this.editLibrary.bind(this);
     this.expandLibrary = this.expandLibrary.bind(this);
     this.removeLibrary = this.removeLibrary.bind(this);
-    this.save = this.save.bind(this);
+    this.handleData = this.handleData.bind(this);
+    this.submit = this.submit.bind(this);
   }
 
   render(): JSX.Element {
     return (
-      <form ref="form" onSubmit={this.save} className="edit-form">
+      <form ref="form" onSubmit={this.submit} className="edit-form">
         { this.props.item && this.props.item.id &&
           <input
             type="hidden"
@@ -68,6 +72,7 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
           disabled={this.props.disabled}
           required={true}
           name="name"
+          ref="name"
           label="Name"
           value={this.props.item && this.props.item.name}
           error={this.props.error}
@@ -81,7 +86,7 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
           value={this.state.protocol}
           ref="protocol"
           onChange={this.handleProtocolChange}
-          description={this.protocolDescription() }
+          description={!this.protocolInstructions() && this.protocolDescription()}
           >
           { this.availableProtocols().map(protocol =>
               <option key={protocol.name} value={protocol.name}>{protocol.label || protocol.name}</option>
@@ -105,9 +110,16 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
             }
           </EditableInput>
         }
+        { this.props.data && this.protocolInstructions() &&
+            <div class="form-group">
+              <label class="control-label">Instructions</label>
+              <Collapsible title={this.protocolDescription()} body={this.protocolInstructions()} />
+            </div>
+        }
         { this.props.data && this.props.data.protocols && this.protocolSettings() && this.protocolSettings().map(setting =>
             <ProtocolFormField
               key={setting.key}
+              ref={setting.key}
               setting={setting}
               disabled={this.props.disabled}
               value={this.props.item && this.props.item.settings && this.props.item.settings[setting.key]}
@@ -123,6 +135,7 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
                   <WithRemoveButton
                     disabled={this.props.disabled}
                     onRemove={() => this.removeLibrary(library)}
+                    ref={library.short_name}
                     >
                     { this.props.data && this.props.data.protocols && this.protocolLibrarySettings() && this.protocolLibrarySettings().length > 0 &&
                       <WithEditButton
@@ -199,12 +212,12 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
             }
           </div>
         }
-        <button
-          className="btn btn-default"
+        <SaveButton
           disabled={this.props.disabled}
-          type="submit">
-          Submit
-        </button>
+          submit={this.submit}
+          text="Submit"
+          form={this.refs}
+        />
       </form>
     );
   }
@@ -235,6 +248,11 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
     }
     const newState = Object.assign({}, this.state, { protocol, parentId, libraries });
     this.setState(newState);
+
+    if (newProps.responseBody && !newProps.fetchError) {
+      clearForm(this.refs);
+    }
+
   }
 
   availableProtocols(props?): ProtocolData[] {
@@ -261,7 +279,7 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
   }
 
   allowsParent(): boolean {
-    if (this.state.protocol && this.props.data.protocols) {
+    if (this.state.protocol && this.props.data && this.props.data.protocols) {
       for (const protocol of this.props.data.protocols) {
         if (protocol.name === this.state.protocol) {
           return !!protocol.child_settings;
@@ -292,7 +310,7 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
   }
 
   protocolSettings() {
-    if (this.state.protocol && this.props.data.protocols) {
+    if (this.state.protocol && this.props.data && this.props.data.protocols) {
       for (const protocol of this.props.data.protocols) {
         if (protocol.name === this.state.protocol) {
           if (this.state.parentId) {
@@ -307,7 +325,7 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
   }
 
   protocolDescription() {
-    if (this.state.protocol && this.props.data.protocols) {
+    if (this.state.protocol && this.props.data && this.props.data.protocols) {
       for (const protocol of this.props.data.protocols) {
         if (protocol.name === this.state.protocol) {
           return protocol.description;
@@ -317,8 +335,19 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
     return "";
   }
 
-  protocolLibrarySettings() {
+  protocolInstructions() {
     if (this.state.protocol && this.props.data.protocols) {
+      for (const protocol of this.props.data.protocols) {
+        if (protocol.name === this.state.protocol) {
+          return protocol.instructions;
+        }
+      }
+    }
+    return "";
+  }
+
+  protocolLibrarySettings() {
+    if (this.state.protocol && this.props.data && this.props.data.protocols) {
       for (const protocol of this.props.data.protocols) {
         if (protocol.name === this.state.protocol) {
           return protocol.library_settings || [];
@@ -329,7 +358,7 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
   }
 
   sitewide() {
-    if (this.state.protocol && this.props.data.protocols) {
+    if (this.state.protocol && this.props.data && this.props.data.protocols) {
       for (const protocol of this.props.data.protocols) {
         if (protocol.name === this.state.protocol) {
           return protocol.sitewide;
@@ -424,16 +453,13 @@ export default class ServiceEditForm<T extends ServicesData> extends React.Compo
     this.setState(newState);
   }
 
-  save(event) {
-    event.preventDefault();
-
-    const data = new (window as any).FormData(this.refs["form"] as any);
-    data.append("libraries", JSON.stringify(this.state.libraries));
-    this.props.editItem(data).then(() => {
-      // If a new service was created, go to its edit page.
-      if (!this.props.item && this.props.responseBody) {
-        window.location.href = this.props.urlBase + "edit/" + this.props.responseBody;
-      }
-    });
+  handleData(data) {
+    return data.append("libraries", JSON.stringify(this.state.libraries));
   }
+
+  submit(event) {
+    event.preventDefault();
+    handleSubmit(this);
+  }
+
 }