+++ src/components/StorageServices.tsx
@@ -19,7 +19,7 @@ function mapStateToProps(state, ownProps) {
   const data = Object.assign({}, state.editor.storageServices && state.editor.storageServices.data || {});
   return {
     data: data,
-    responseBody: state.editor.storageServices && state.editor.storageServices.responseBody,
+    responseBody: state.editor.storageServices && state.editor.storageServices.successMessage,
     fetchError: state.editor.storageServices.fetchError,
     isFetching: state.editor.storageServices.isFetching || state.editor.storageServices.isEditing
   };