+++ src/interfaces.ts
@@ -185,6 +185,7 @@ export interface ProtocolData {
   name: string;
   label?: string;
   description?: string;
+  instructions?: string;
   sitewide?: boolean;
   supports_registration?: boolean;
   settings: SettingData[];